requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
DateTime==5.0
Django==4.1.6
django-rest-swagger==2.2.0
djangorestframework==3.14.0
docx==0.2.4
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
openapi-codec==1.3.2
pdfminer==20191125
pdfminer.six==20221105
Pillow==9.4.0
pycparser==2.21
pycryptodome==3.16.0
PyMySQL==1.0.2
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-docx==0.8.11
pytz==2022.7.1
requests==2.28.2
simplejson==3.18.3
sqlparse==0.4.3
tk==0.1.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
zope.interface==5.5.2
